agi/tests/test_event_loop.py
import unittest
from unittest.mock import patch, MagicMock
from concurrent.futures import ProcessPoolExecutor
from multiprocessing import Manager
from agi import (
        chat,
        #main,
        # handle_user_input,
        #handle_boredom  # Adjust import as necessary
        )
from agi.event_loop import (
        handle_user_input,
        handle_boredom  # Adjust import as necessary

        )

class TestEventLoop(unittest.TestCase):
    def setUp(self):
        # Set up a Manager to simulate in tests
        self.manager = Manager()
        self.boredom_counter = self.manager.Value('i', 0)
        self.input_event = self.manager.Event()

    @patch('your_module.chat.user_input', return_value='!multi line1 line2 !end')
    def test_handle_user_input(self, mock_input):
        """Test that user input properly triggers the input event."""
        with patch('your_module.chat.get_db') as mock_get_db:
            mock_get_db.return_value = MagicMock()
            # Assume user_input is a generator; simulate stepping through inputs
            mock_input.return_value = iter(['!multi', 'line1', 'line2', '!end line1 line2'])
            inputs = list(chat.user_input())  # Would normally loop forever; here we simulate and convert to list
            self.assertEqual(inputs, ['line1\nline2'])  # Check if multi-line input was handled correctly
    
    @patch('time.sleep', MagicMock())  # Mock sleep to speed up the test
    def test_handle_boredom(self):
        """Test that boredom is reset on user input."""
        # Trigger the input event as if user has interacted
        self.input_event.set()
        handle_boredom(self.boredom_counter, self.input_event)
        self.assertEqual(self.boredom_counter.value, 0)
        self.assertFalse(self.input_event.is_set())  # Event should be cleared after handling

# Additional setup for running the tests if not using a test runner like pytest
if __name__ == '__main__':
    unittest.main()
P
